+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Upcoming
+## 3.6.5
 * Wither
     * Bonus health regen is now reduced by 40% if Wither took damage in the last 3 seconds
 * Dragon